move Utils to his own file
